requirements.txt
# VoiceMessage MCP Server Requirements
# Install with: pip install -r requirements.txt

# MCP Server
mcp>=1.0.0

# Text-to-Speech (Kokoro ONNX)
kokoro-onnx>=0.2.0

# Audio processing
numpy>=1.24.0
soundfile>=0.12.0

# Environment variables
python-dotenv>=1.0.0

# Cloudflare R2 storage
boto3>=1.28.0

# Twilio API
twilio>=8.10.0